buyv_backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0

# Database & Storage
SQLAlchemy==2.0.32
PyMySQL==1.1.0
psycopg2-binary==2.9.10
cloudinary==1.44.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
slowapi==0.1.9

# Payments
stripe==11.4.0

# Firebase
firebase-admin==6.5.0

# Utilities
pydantic==2.9.1
python-multipart==0.0.20
